Move pipeline file to index.
